+++ src/RoundSwitch/index.d.ts
@@ -1,0 +1,4 @@
+import type { SvelteComponentTyped } from 'svelte'
+export class RoundSwitch extends SvelteComponentTyped<{
+	id?:string
+}> {}