+++ src/RoundSwitch/index.svelte
@@ -1,4 +1,4 @@
-<script lang="ts">
+<script>
 export let id = ''
 </script>
 
@@ -7,39 +7,34 @@ export let id = ''
 	<span class="slider round"></span>
 </label>
 
-<style lang="scss">
+<style>
 .Switch {
 	position: relative;
 	display: inline-block;
 	width: 60px;
 	height: 34px;
-	input {
+}
+	.Switch input {
 		display: none;
-		&:checked {
-			+ .slider {
+	}
+			.Switch input:checked + .slider {
 				background-color: #3EBBC0;
-				&:before {
+			}
+				.Switch input:checked + .slider:before {
 					-webkit-transform: translateX(26px);
 					-ms-transform: translateX(26px);
 					transform: translateX(26px);
 				}
-			}
-		}
-		&:focused {
-			+ .slider {
+			.Switch input:focused + .slider {
 				box-shadow: 0 0 1px #2196F3;
 			}
-		}
-		+ .slider {
-			&.round {
+			.Switch input + .slider.round {
 				border-radius: 34px;
-				&:before {
+			}
+				.Switch input + .slider.round:before {
 					border-radius: 50%;
 				}
-			}
-		}
-	}
-	.slider {
+	.Switch .slider {
 		position: absolute;
 		cursor: pointer;
 		top: 0;
@@ -49,7 +44,8 @@ export let id = ''
 		background-color: #ccc;
 		-webkit-transition: .4s;
 		transition: .4s;
-		&:before {
+	}
+		.Switch .slider:before {
 			position: absolute;
 			content: "";
 			height: 26px;
@@ -60,6 +56,4 @@ export let id = ''
 			-webkit-transition: .4s;
 			transition: .4s;
 		}
-	}
-}
 </style>