+++ src/index.ts
@@ -1,2 +1,0 @@
-import 'svelte'
-export * as RoundSwitch from './RoundSwitch.svelte'